+++ typescript/cli/src/tests/ethereum/warp/warp-apply-2.e2e-test.ts
@@ -4,13 +4,11 @@ import { Wallet, ethers } from 'ethers';
 import { ChainAddresses } from '@hyperlane-xyz/registry';
 import {
   ChainMetadata,
-  HookType,
   HypTokenRouterConfig,
   HypTokenRouterConfigMailboxOptionalSchema,
   TokenType,
   WarpCoreConfig,
   WarpRouteDeployConfig,
-  normalizeConfig,
   randomAddress,
 } from '@hyperlane-xyz/sdk';
 import {
@@ -30,7 +28,6 @@ import {
   hyperlaneWarpApplyRaw,
   hyperlaneWarpDeploy,
   readWarpConfig,
-  updateOwner,
 } from '../commands/warp.js';
 import {
   ANVIL_DEPLOYER_ADDRESS,
@@ -73,166 +70,6 @@ describe('hyperlane warp apply owner update tests', async function () {
 
   beforeEach(async function () {
     await hyperlaneWarpDeploy(WARP_CONFIG_PATH_2, WARP_DEPLOY_2_ID);
-  });
-
-  it('should burn owner address', async function () {
-    const warpConfigPath = `${TEMP_PATH}/warp-route-deployment-2.yaml`;
-    await updateOwner(
-      E2E_TEST_BURN_ADDRESS,
-      CHAIN_NAME_2,
-      warpConfigPath,
-      WARP_CORE_CONFIG_PATH_2,
-    );
-    const updatedWarpDeployConfig = await readWarpConfig(
-      CHAIN_NAME_2,
-      WARP_CORE_CONFIG_PATH_2,
-      warpConfigPath,
-    );
-    expect(updatedWarpDeployConfig.anvil2.owner).to.equal(
-      E2E_TEST_BURN_ADDRESS,
-    );
-  });
-
-  it('should not update the same owner', async () => {
-    const warpConfigPath = `${TEMP_PATH}/warp-route-deployment-2.yaml`;
-    await updateOwner(
-      E2E_TEST_BURN_ADDRESS,
-      CHAIN_NAME_2,
-      warpConfigPath,
-      WARP_CORE_CONFIG_PATH_2,
-    );
-    const { stdout } = await updateOwner(
-      E2E_TEST_BURN_ADDRESS,
-      CHAIN_NAME_2,
-      warpConfigPath,
-      WARP_CORE_CONFIG_PATH_2,
-    );
-    expect(stdout).to.include(
-      'Warp config is the same as target. No updates needed.',
-    );
-  });
-
-  it('should update the owner of both the warp token and the proxy admin', async () => {
-    const warpConfigPath = `${TEMP_PATH}/warp-route-deploy-config-2.yaml`;
-
-    const warpConfig: WarpRouteDeployConfig = readYamlOrJson(
-      WARP_CONFIG_PATH_EXAMPLE,
-    );
-
-    // Set to undefined if it was defined in the config
-    warpConfig.anvil1.proxyAdmin = undefined;
-    warpConfig.anvil1.owner = E2E_TEST_BURN_ADDRESS;
-    const anvil2Config = { anvil2: { ...warpConfig.anvil1 } };
-    writeYamlOrJson(warpConfigPath, anvil2Config);
-
-    await hyperlaneWarpApply(
-      warpConfigPath,
-      WARP_CORE_CONFIG_PATH_2,
-      undefined,
-      WARP_DEPLOY_2_ID,
-    );
-
-    const updatedWarpDeployConfig1 = await readWarpConfig(
-      CHAIN_NAME_2,
-      WARP_CORE_CONFIG_PATH_2,
-      warpConfigPath,
-    );
-
-    expect(updatedWarpDeployConfig1.anvil2.owner).to.eq(E2E_TEST_BURN_ADDRESS);
-    expect(updatedWarpDeployConfig1.anvil2.proxyAdmin?.owner).to.eq(
-      E2E_TEST_BURN_ADDRESS,
-    );
-  });
-
-  it('should update only the owner of the warp token if the proxy admin config is specified', async () => {
-    const warpConfigPath = `${TEMP_PATH}/warp-route-deploy-config-2.yaml`;
-
-    const warpConfig: WarpRouteDeployConfig = readYamlOrJson(
-      WARP_CONFIG_PATH_EXAMPLE,
-    );
-
-    // Explicitly set it to the deployer address if it was not defined
-    warpConfig.anvil1.proxyAdmin = { owner: ANVIL_DEPLOYER_ADDRESS };
-    warpConfig.anvil1.owner = E2E_TEST_BURN_ADDRESS;
-    const anvil2Config = { anvil2: { ...warpConfig.anvil1 } };
-    writeYamlOrJson(warpConfigPath, anvil2Config);
-
-    await hyperlaneWarpApply(warpConfigPath, WARP_CORE_CONFIG_PATH_2);
-
-    const updatedWarpDeployConfig1 = await readWarpConfig(
-      CHAIN_NAME_2,
-      WARP_CORE_CONFIG_PATH_2,
-      warpConfigPath,
-    );
-
-    expect(updatedWarpDeployConfig1.anvil2.owner).to.eq(E2E_TEST_BURN_ADDRESS);
-    expect(updatedWarpDeployConfig1.anvil2.proxyAdmin?.owner).to.eq(
-      ANVIL_DEPLOYER_ADDRESS,
-    );
-  });
-
-  it('should update only the owner of the proxy admin if the proxy admin config is specified', async () => {
-    const warpConfigPath = `${TEMP_PATH}/warp-route-deploy-config-2.yaml`;
-
-    const warpConfig: WarpRouteDeployConfig = readYamlOrJson(
-      WARP_CONFIG_PATH_EXAMPLE,
-    );
-
-    warpConfig.anvil1.proxyAdmin = { owner: E2E_TEST_BURN_ADDRESS };
-    const anvil2Config = { anvil2: { ...warpConfig.anvil1 } };
-    writeYamlOrJson(warpConfigPath, anvil2Config);
-
-    await hyperlaneWarpApply(warpConfigPath, WARP_CORE_CONFIG_PATH_2);
-
-    const updatedWarpDeployConfig1 = await readWarpConfig(
-      CHAIN_NAME_2,
-      WARP_CORE_CONFIG_PATH_2,
-      warpConfigPath,
-    );
-
-    expect(updatedWarpDeployConfig1.anvil2.owner).to.eq(ANVIL_DEPLOYER_ADDRESS);
-    expect(updatedWarpDeployConfig1.anvil2.proxyAdmin?.owner).to.eq(
-      E2E_TEST_BURN_ADDRESS,
-    );
-  });
-
-  it('should update hook configuration', async () => {
-    const warpDeployPath = `${TEMP_PATH}/warp-route-deployment-2.yaml`;
-
-    // First read the existing config
-    const warpDeployConfig = await readWarpConfig(
-      CHAIN_NAME_2,
-      WARP_CORE_CONFIG_PATH_2,
-      warpDeployPath,
-    );
-
-    // Update with a new hook config
-    const owner = randomAddress();
-    warpDeployConfig[CHAIN_NAME_2].hook = {
-      type: HookType.PROTOCOL_FEE,
-      beneficiary: owner,
-      maxProtocolFee: '1000000',
-      protocolFee: '100000',
-      owner,
-    };
-
-    // Write the updated config
-    await writeYamlOrJson(warpDeployPath, warpDeployConfig);
-
-    // Apply the changes
-    await hyperlaneWarpApply(warpDeployPath, WARP_CORE_CONFIG_PATH_2);
-
-    // Read back the config to verify changes
-    const updatedConfig = await readWarpConfig(
-      CHAIN_NAME_2,
-      WARP_CORE_CONFIG_PATH_2,
-      warpDeployPath,
-    );
-
-    // Verify the hook was updated with all properties
-    expect(normalizeConfig(updatedConfig[CHAIN_NAME_2].hook)).to.deep.equal(
-      normalizeConfig(warpDeployConfig[CHAIN_NAME_2].hook),
-    );
   });
 
   it('should extend a warp route with a custom warp route id', async () => {
